+++ test/integration/scss-fixtures/webpack-error/mock.js
@@ -2,6 +2,6 @@ let originalLoader
 const M = require('module')
 originalLoader = M._load
 M._load = function hookedLoader(request, parent, isMain) {
-  if (request === 'node-sass') request = 'node-sass-begone'
+  if (request === 'sass') request = 'sass-begone'
   return originalLoader(request, parent, isMain)
 }
